ship started

+++ Bullet.java
@@ -2,14 +2,12 @@ public class Bullet extends SpaceObject {
 	private int timeLeft = 1000; // Muna að stilla hraðann!
 
 	public Bullet(double x, double y, int dir) {
-		// SpaceObject(double x, double y, double w, double h, double spd, int dir, int rot)
-		super(x, y, 0.05 * Math.cos(Math.toRadians(dir)), 0.05 * Math.sin(Math.toRadians(dir)), 0.001, dir, dir);
+		// SpaceObject(double x, double y, double w, double h, double spd, double dir, int rot)
+		super(x, y, 0.05 * Math.cos(Math.toRadians(dir)), 0.05 * Math.sin(Math.toRadians(dir)), 0.001, (double)dir, dir);
 	}
 	
 	public void draw() {
 		if(timeLeft > 0) {
-
-
 			StdDraw.line(this.x, this.y, this.x+this.w, this.y+this.h);
 			timeLeft--;
 		}
